Moved utils to shared folder
